pyproject.toml
[project]
name = "pymaskinporten"
version = "1.1.0"
description = "Fetch Maskinporten Token"
readme = "README.md"
license = "Apache-2.0"
authors = [{ name = "Novica Nakov", email = "nnovica@gmail.com" }]

requires-python = ">=3.13"
dependencies = [
  "azure-identity>=1.25.2",
  "azure-keyvault-secrets>=4.10.0",
  "flask>=3.1.3",
  "httpx>=0.28.1",
  "jwcrypto>=1.5.6",
  "pydantic>=2.12.5",
  "pyjwt>=2.11.0",
  "python-dotenv>=1.2.2",
]

[project.scripts]
maskinporten_web = "pymaskinporten.app:main"
pymaskinporten = "pymaskinporten:main"

[dependency-groups]
dev = [
  "pytest>=9.0.0",
  "ruff>=0.6.1",
  "ty>=0.0.21",
]

[build-system]
requires = ["uv_build>=0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "pymaskinporten"
module-root = "src"